refactor(tidy): rename agent marketplace boundary to hub
