+++ backend/library/__init__.py
@@ -1,24 +1,17 @@
-"""Library domain — file-backed + DB-backed asset library.
-
-TENTATIVE DOMAIN. Target architecture §5.5 marks library as deferred;
-boundaries may shift once an audit decides whether this is a real
-standalone domain or a chat-side projection. This package is a temporary
-home to get library code out of backend/web/services/ without settling
-the domain question.
+"""Library domain — product-side asset/template/skill library.
 
 IN:
     - paths.py (LIBRARY_DIR)
     - service.py (library CRUD over files and recipe repo)
-    - marketplace_client.py (HTTP client for Mycel Hub)
 
 OUT:
+    - Mycel-Hub integration (backend/hub/)
     - Sandbox recipe definitions (sandbox/recipes.py)
     - Agent user registration (backend/identity/ + backend/threads/)
-    - Snapshot installation (backend/agent_marketplace/snapshot_install.py)
 
 Dependencies:
     top-level: sandbox/, storage/, config/
-    backend:   identity/, sandboxes/, agent_marketplace/
+    backend:   identity/, sandboxes/
 
 See program/doc/core/backend-package-dependencies-2026-04-20.md §5.6.
 """
