+++ backend/web/routers/marketplace.py
@@ -5,6 +5,7 @@ from typing import Annotated, Any
 
 from fastapi import APIRouter, Depends, HTTPException, Request
 
+from backend.hub import client as marketplace_client
 from backend.identity.profile import get_profile
 from backend.web.core.dependencies import get_current_user_id
 from backend.web.models.marketplace import (
@@ -13,7 +14,6 @@ from backend.web.models.marketplace import (
     PublishAgentUserToMarketplaceRequest,
     UpgradeFromMarketplaceRequest,
 )
-from backend.library import marketplace_client
 
 router = APIRouter(prefix="/api/marketplace", tags=["marketplace"])
